[backend] Retrieve entities from container for entity
