+++ opencti-platform/opencti-graphql/src/domain/stixDomainObject.js
@@ -10,11 +10,12 @@ import {
   timeSeriesEntities,
   updateAttribute,
 } from '../database/middleware';
-import { listEntities, storeLoadById } from '../database/middleware-loader';
+import { internalFindByIds, internalLoadById, listEntities, listEntitiesPaginated, storeLoadById } from '../database/middleware-loader';
 import { elCount, elFindByIds } from '../database/engine';
 import { workToExportFile } from './work';
 import { FunctionalError, UnsupportedError } from '../config/errors';
 import {
+  buildPagination,
   isEmptyField,
   isNotEmptyField,
   READ_INDEX_INFERRED_ENTITIES,
@@ -30,7 +31,7 @@ import {
   isStixDomainObjectLocation, isStixDomainObjectThreatActor,
 } from '../schema/stixDomainObject';
 import { ABSTRACT_STIX_CYBER_OBSERVABLE, ABSTRACT_STIX_DOMAIN_OBJECT, buildRefRelationKey, } from '../schema/general';
-import { RELATION_CREATED_BY, RELATION_OBJECT_ASSIGNEE, } from '../schema/stixRefRelationship';
+import { RELATION_CREATED_BY, RELATION_OBJECT, RELATION_OBJECT_ASSIGNEE, } from '../schema/stixRefRelationship';
 import { askEntityExport, askListExport, exportTransformFilters } from './stix';
 import { RELATION_BASED_ON } from '../schema/stixCoreRelationship';
 import { now, utcDate } from '../utils/format';
@@ -251,4 +252,27 @@ export const stixDomainObjectEditContext = async (context, user, stixDomainObjec
     return notify(BUS_TOPICS[ABSTRACT_STIX_DOMAIN_OBJECT].CONTEXT_TOPIC, stixDomainObject, user);
   });
 };
+// endregion
+
+// region container
+export const stixDomainObjectsRelatedObjectsFromContainer = async (context, user, args) => {
+  const { fromId, entityType, containerType } = args;
+  // Retrieve entity type
+  const entity = await internalLoadById(context, user, fromId);
+  // Get rel_object.internal_id from entity
+  const containerIds = entity['rel_object.internal_id'];
+  // Filter on container type
+  const containers = await internalFindByIds(context, user, containerIds, { baseData: true })
+  const containerTypeIds = containers.filter((o) => o.entity_type === containerType).map((r) => r.internal_id);
+  if (containerTypeIds.length === 0) {
+    return buildPagination(0, null, [], 0);
+  }
+  // Retrieve all entities with rel_object.internal_id contains at least one of containerTypeIds
+  const opts = {
+    filters: [
+      { key: buildRefRelationKey(RELATION_OBJECT), values: containerTypeIds, operator: 'match' },
+    ],
+  };
+  return listEntitiesPaginated(context, user, [entityType], opts);
+}
 // endregion
