+++ opencti-platform/opencti-graphql/src/modules/threatActorIndividual/threatActorIndividual-types.ts
@@ -20,8 +20,6 @@ export interface BasicStoreEntityThreatActorIndividual extends BasicStoreEntity 
   secondary_motivations: string[]
   personal_motivations: string[]
   eye_color: string
-  // internal
-  ['rel_object.internal_id']: string[]
 }
 
 export interface StoreEntityThreatActorIndividual extends StoreEntity {
