+++ opencti-platform/opencti-graphql/src/modules/threatActorIndividual/threatActorIndividual.graphql
@@ -85,7 +85,7 @@ type ThreatActorIndividual implements BasicObject & StixObject & StixCoreObject 
     filters: [StixCoreRelationshipsFiltering]
     filterMode: FilterMode
   ): [Distribution] @auth(for: [KNOWLEDGE, EXPLORE])
-  # StixDomainObject
+  stixCoreObjectFromContainer(fromId: String!, entityType: String!, containerType: String!): StixCoreObjectConnection @auth(for: [KNOWLEDGE])# StixDomainObject
   revoked: Boolean!
   confidence: Int
   lang: String
@@ -209,7 +209,6 @@ type Query {
     toStix: Boolean
   ): ThreatActorIndividualConnection @auth(for: [KNOWLEDGE])
   threatActorIndividualContainsStixObjectOrStixRelationship(id: String!, stixObjectOrStixRelationshipId: String!): Boolean @auth(for: [KNOWLEDGE])
-  stixCoreObjectFromContainer(fromId: String!, entityType: String!, containerType: String!): StixCoreObjectConnection @auth(for: [KNOWLEDGE])
 }
 
 type Mutation {
